+++ domain/src/main/java/com/albertmiro/domain/models/Pilot.kt
@@ -3,5 +3,5 @@ package com.albertmiro.domain.models
 data class Pilot(
     val name: String,
     val avatar: String,
-    val rating: Double
+    val rating: Float
 )
