+++ domain/src/main/java/com/albertmiro/domain/usecases/GetTripDetails.kt
@@ -3,6 +3,6 @@ package com.albertmiro.domain.usecases
 import com.albertmiro.domain.repository.TripsRepository
 
 class GetTripDetails(val tripsRepository: TripsRepository) {
-    operator fun invoke(id: Int) =
+    operator fun invoke(id: Long) =
         tripsRepository.getTripDetail(id)
 }
